Move ListBinarySerializer tests to its own file
