+++ SAGESharpTests/SAGESharpTests.csproj
@@ -89,6 +89,7 @@
     <Compile Include="IO\DefaultBinarySerializerTests.cs" />
     <Compile Include="IO\DefaultPropertyBinarySerializerFactoryTests.cs" />
     <Compile Include="IO\DefaultPropertyBinarySerializerTests.cs" />
+    <Compile Include="IO\ListBinarySerializerTests.cs" />
     <Compile Include="IO\PrimitiveBinarySerializerTests.cs" />
     <Compile Include="IO\StringBinarySerializerTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
